No need for a seperate file
